+++ examples/functions/misc.js
@@ -1,0 +1,2 @@
+global.Deno = require('../..');
+require('./misc.ts');