+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning 2.0.0](http://semver.org/).
 
+##  [0.0.4] - May 17th, 2019
+
+*   Function [chdir()](./docs/functions.md#chdir) added.
+*   Function [cwd()](./docs/functions.md#cwd) added.
+*   Function [exit()](./docs/functions.md#exit) added.
+*   Function [kill()](./docs/functions.md#kill) added.
+*   Function [mkdir()](./docs/functions.md#mkdir) added.
+*   Function [mkdirSync()](./docs/functions.md#mkdirsync) added.
+*   Function [remove()](./docs/functions.md#remove) added.
+*   Function [removeSync()](./docs/functions.md#removesync) added.
+
 ##  [0.0.3] - May 12th, 2019
 
 *   [datetime](./datetime.js) implemented.